# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.AuthRuleResponse:
    properties:
      config:
        additionalProperties: true
        type: object
      enabled:
        type: boolean
      type:
        type: string
    type: object
...
  models.BucketSettingsResponse:
    properties:
      allow_overwrite:
        type: boolean
      allowed_extensions:
        items:
          type: string
        type: array
      allowed_mime_types:
        items:
          type: string
        type: array
      blocked_extensions:
        items:
          type: string
        type: array
      blocked_mime_types:
        items:
          type: string
        type: array
      encryption:
        type: boolean
      max_file_size:
        type: integer
      max_files_per_bucket:
        type: integer
      max_total_size:
        type: integer
      public_read:
        type: boolean
      require_content_type:
        type: boolean
      versioning:
        type: boolean
    type: object
...
  models.MasterSetupRequest:
    properties:
      admin_email:
        type: string
      admin_password:
        minLength: 6
        type: string
      admin_username:
        maxLength: 50
        minLength: 3
        type: string
      default_auth_rule:
        $ref: '#/definitions/models.AuthRuleResponse'
      default_settings:
        $ref: '#/definitions/models.BucketSettingsResponse'
      jwt_secret:
        type: string
      max_storage:
        minimum: 1
        type: integer
      storage_path:
        type: string
      system_name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - admin_email
    - admin_password
    - admin_username
    - storage_path
    - system_name
    type: object
...
  models.NodeSetupRequest:
    properties:
      master_api_key:
        type: string
      master_url:
        type: string
      max_storage:
        minimum: 1
        type: integer
      node_api_key:
        minLength: 10
        type: string
      node_name:
        maxLength: 100
        minLength: 3
        type: string
      storage_path:
        type: string
    required:
    - master_api_key
    - master_url
    - node_api_key
    - node_name
    - storage_path
    type: object
...
  models.StorageNodeResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_healthy:
        type: boolean
      last_ping:
        type: string
      max_storage:
        type: integer
      name:
        type: string
      priority:
        type: integer
      updated_at:
        type: string
      url:
        type: string
      used_storage:
        type: integer
    type: object
  models.SystemInfoResponse:
    properties:
      free_storage:
        type: integer
      is_healthy:
        type: boolean
      last_checked:
        type: string
      max_storage:
        type: integer
      node_count:
        type: integer
      setup_type:
        type: string
      storage_path:
        type: string
      system_name:
        type: string
      used_storage:
        type: integer
      version:
        type: string
    type: object
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  setup.CheckSetupResponse:
    properties:
      is_setup:
        type: boolean
      message:
        type: string
      node_name:
        type: string
      setup_type:
        type: string
    type: object
  setup.MasterSetupResponse:
    properties:
      admin_user:
        $ref: '#/definitions/models.UserResponse'
      config:
        additionalProperties: true
        type: object
      message:
        type: string
      success:
        type: boolean
    type: object
  setup.NodeSetupResponse:
    properties:
      config:
        additionalProperties: true
        type: object
      message:
        type: string
      node:
        $ref: '#/definitions/models.StorageNodeResponse'
      success:
        type: boolean
    type: object
  user.ChangePasswordCommand:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
      user_id:
        type: string
    required:
    - new_password
    - old_password
    type: object
  user.ChangePasswordResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  user.GetUserResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  user.ListUsersResponse:
    properties:
      limit:
        type: integer
      message:
        type: string
      page:
        type: integer
      success:
        type: boolean
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.UserResponse'
        type: array
    type: object
  user.LoginCommand:
    properties:
      email:
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  user.LoginResponse:
    properties:
      expires_in:
        type: integer
      message:
        type: string
      refresh_token:
        type: string
      success:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  user.LogoutResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  user.RegisterCommand:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      role:
        enum:
        - viewer
        - editor
        - manager
        - admin
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  user.RegisterResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
host: localhost:8080
info:
  contact:
    email: support@shbucket.local
    name: SHBucket Support
  description: SHBucket is a distributed object storage system similar to AWS S3
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SHBucket API
  version: 2.0.0
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change the password for the authenticated user
      parameters:
      - description: Password change details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.ChangePasswordCommand'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/user.ChangePasswordResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Change password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, returns JWT token for
        subsequent requests
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/user.LoginCommand'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/user.LoginResponse'
        "400":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user and invalidate session token
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/user.LogoutResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: User logout
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.RegisterCommand'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/user.RegisterResponse'
        "400":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User registration
      tags:
      - auth
...
  /setup/info:
    get:
      consumes:
      - application/json
      description: Retrieve system status and information after setup
      produces:
      - application/json
      responses:
        "200":
          description: System information
          schema:
            $ref: '#/definitions/models.SystemInfoResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get system information
      tags:
      - setup
  /setup/master:
    post:
      consumes:
      - application/json
      description: Initialize the system as a master node with database and admin
        user
      parameters:
      - description: Master setup configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MasterSetupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Master setup successful
          schema:
            $ref: '#/definitions/setup.MasterSetupResponse'
        "400":
          description: Invalid request or setup failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Setup master node
      tags:
      - setup
  /setup/node:
    post:
      consumes:
      - application/json
      description: Initialize the system as a storage node connected to a master
      parameters:
      - description: Node setup configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.NodeSetupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Node setup successful
          schema:
            $ref: '#/definitions/setup.NodeSetupResponse'
        "400":
          description: Invalid request or setup failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Setup node
      tags:
      - setup
  /setup/status:
    get:
      consumes:
      - application/json
      description: Check if the system has been set up and configured
      produces:
      - application/json
      responses:
        "200":
          description: Setup status information
          schema:
            $ref: '#/definitions/setup.CheckSetupResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check setup status
      tags:
      - setup
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all users (admin only)
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Include user buckets
        in: query
        name: include_buckets
        type: boolean
      - description: Include user sessions
        in: query
        name: include_sessions
        type: boolean
      - description: Include all related data
        in: query
        name: include_all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            $ref: '#/definitions/user.ListUsersResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List users
      tags:
      - users
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get information about a specific user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/user.GetUserResponse'
        "400":
          description: Invalid user ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get user by ID
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: API Key for authentication
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"